move ner_re-pipelinr related code from utils.document to ner_re_pipeline.py
