+++ src/utils/document.py
@@ -215,156 +215,6 @@ def remove_overlapping_entities(
     return new_doc
 
 
-def clear_annotation_layers(
-    doc: D, layer_names: List[str], predictions: bool = False, inplace: bool = False
-) -> D:
-    if not inplace:
-        doc = type(doc).fromdict(doc.asdict())
-    for layer_name in layer_names:
-        if predictions:
-            doc[layer_name].predictions.clear()
-        else:
-            doc[layer_name].clear()
-    return doc
-
-
-def move_annotations_from_predictions(doc: D, layer_names: List[str], inplace: bool = False) -> D:
-    if not inplace:
-        doc = type(doc).fromdict(doc.asdict())
-    for layer_name in layer_names:
-        annotations = list(doc[layer_name].predictions)
-        # remove any previous annotations
-        doc[layer_name].clear()
-        # each annotation can be attached to just one annotation container, so we need to clear the predictions
-        doc[layer_name].predictions.clear()
-        doc[layer_name].extend(annotations)
-    return doc
-
-
-def move_annotations_to_predictions(doc: D, layer_names: List[str], inplace: bool = False) -> D:
-    if not inplace:
-        doc = type(doc).fromdict(doc.asdict())
-    for layer_name in layer_names:
-        annotations = list(doc[layer_name])
-        # each annotation can be attached to just one annotation container, so we need to clear the layer
-        doc[layer_name].clear()
-        # remove any previous annotations
-        doc[layer_name].predictions.clear()
-        doc[layer_name].predictions.extend(annotations)
-    return doc
-
-
-def add_annotations_from_other_documents(
-    docs: List[D],
-    other_docs: List[Document],
-    layer_names: List[str],
-    from_predictions: bool = False,
-    to_predictions: bool = False,
-    clear_before: bool = True,
-    inplace: bool = False,
-) -> List[D]:
-    prepared_documents = []
-    for i, doc in enumerate(docs):
-        if not inplace:
-            doc = type(doc).fromdict(doc.asdict())
-        other_doc = other_docs[i]
-        # copy to not modify the input
-        other_doc = type(other_doc).fromdict(other_doc.asdict())
-
-        for layer_name in layer_names:
-            if clear_before:
-                doc[layer_name].clear()
-            other_layer = other_doc[layer_name]
-            if from_predictions:
-                other_layer = other_layer.predictions
-            other_annotations = list(other_layer)
-            other_layer.clear()
-            if to_predictions:
-                doc[layer_name].predictions.extend(other_annotations)
-            else:
-                doc[layer_name].extend(other_annotations)
-        prepared_documents.append(doc)
-    return prepared_documents
-
-
-def trim_spans(
-    document: D,
-    span_layer: str,
-    text_field: str = "text",
-    relation_layer: Optional[str] = None,
-    skip_empty: bool = True,
-    inplace: bool = False,
-) -> D:
-    """Remove the whitespace at the beginning and end of span annotations. If a relation layer is
-    given, the relations will be updated to point to the new spans.
-
-    Args:
-        document: The document to trim its span annotations.
-        span_layer: The name of the span layer to trim.
-        text_field: The name of the text field in the document.
-        relation_layer: The name of the relation layer to update. If None, the relations will not be updated.
-        skip_empty: If True, empty spans will be skipped. Otherwise, an error will be raised.
-        copy: If True, the document will be copied before trimming.
-
-    Returns:
-        The document with trimmed spans.
-    """
-    if not inplace:
-        document = type(document).fromdict(document.asdict())
-
-    spans: AnnotationList[LabeledSpan] = document[span_layer]
-    old2new_spans = {}
-    text = getattr(document, text_field)
-    for span in spans:
-        span_text = text[span.start : span.end]
-        new_start = span.start + len(span_text) - len(span_text.lstrip())
-        new_end = span.end - len(span_text) + len(span_text.rstrip())
-        # if the new span is empty and skip_empty is True, skip it
-        if new_start < new_end or not skip_empty:
-            # if skip_empty is False and the new span is empty, log a warning and create a new zero-width span
-            # by using the old start position
-            if new_start >= new_end:
-                logger.warning(
-                    f"Span {span} is empty after trimming. {'Skipping it.' if skip_empty else ''}"
-                )
-                new_start = span.start
-                new_end = span.start
-            new_span = LabeledSpan(
-                start=new_start,
-                end=new_end,
-                label=span.label,
-                score=span.score,
-            )
-            old2new_spans[span] = new_span
-
-    spans.clear()
-    spans.extend(old2new_spans.values())
-
-    if relation_layer is not None:
-        relations: AnnotationList[BinaryRelation] = document[relation_layer]
-        new_relations = []
-        for relation in relations:
-            if relation.head not in old2new_spans or relation.tail not in old2new_spans:
-                logger.warning(
-                    f"Relation {relation} is removed because one of its spans was removed."
-                )
-                continue
-            head = old2new_spans[relation.head]
-            tail = old2new_spans[relation.tail]
-            new_relations.append(
-                BinaryRelation(
-                    head=head,
-                    tail=tail,
-                    label=relation.label,
-                    score=relation.score,
-                )
-            )
-        relations.clear()
-        relations.extend(new_relations)
-
-    return document
-
-
 def add_reversed_relations(
     document: D,
     relation_layer: str = "relations",
